Set all cookies to SameSite=Lax
The login flow is broken if either the OAuth state cookie or session
cookie are SameSite=Strict _and_ the user was not already logged in to
GitHub when they began the login flow for the developer portal. If the
user is already logged in to GitHub, everything works fine even if both
cookies are set to SameSite=Strict.

This bug previously went unnoticed as I was not logging out of GitHub
for every login test. Further investigation is needed to see if we can
set the cookies back to SameSite=strict (again) without breaking the
login flow.