Add initial support for app submissions
There is now support for validated, authenticated new app submissions.
since this is only the initial implementation, error handling and
reporting is poor, proper rate limits aren't in place to prevent abuse,
and cleanup of temporary files isn't perfect. However, all submitted
apps are checked to be in the proper format (to a degree) and their
metadata is saved in the database for later approval by an admin.

Closes #7