Re-implement registration page in Angular
The usable_email_cache but-not-actually-a-cache table has been purged in
favor of a repository pattern via functions in the `data` module. It can
be added back in the future under this module as an actual cache, but
route handlers should never access it directly, only through the `data`
module abstraction.