Always assign apps & updates to individual reviewers
This commit's purpose is twofold. The first is to always require manual
review for new apps because it's impossible to foresee all reasons an
app should be rejected and it's desirable to always manually check for
e.g. trademark violations.

The second purpose is to assign all app submissions to individual
reviewers instead of allowing every reviewer to review every app. This
prevents conflicts and duplication of effort if multiple reviewers
decide to review the same app at the same time and sets a single line of
communication for the submitter if they have review questions.

Note that if a reviewer themself submits an app, it is possible that
they are assigned to review their own app. This behavior isn't desirable
for obvious reasons, but it's a low priority to fix since reviewers are
already trusted to some degree.