Publish update files before beginning transaction
Similar to how publish_app.go works, attempt to publish the app update
to the repository server before touching the database. This decouples
the file publishing step from the database modification step, allowing
the latter to be extracted and abstracted in the future.

This change also fixes a bug where if file publishing failed, the
current database transaction wouldn't be rolled back (which is another
reason to separate the publishing steps).