Encapsulate database with new `data.DB` interface
Route handlers no longer make SQL queries directly. They instead take a
`data.DB` interface and call the appropriate methods on it.

This change is immensely useful for many reasons:

- similar queries can be easily deduplicated by making another method on
  the database struct
- it's straightforward to swap out the database implementation if needed
- we can now create a mock database for testing purposes
- it's impossible for non-database logic to interfere with database
  logic by, for example, returning early in the middle of a transaction
- route handlers are now much simpler and easier to read

Currently the default SQLite implementation of `data.DB` propogates
package-specific errors to the caller, requiring them to handle
implementation details. Returning errors generic over the implementation
is left for a future commit.

Some methods also expose underlying database fields such as
`issue_group_id` which needn't ever be visible to the caller. However,
in the interest of scoping this commit to one thing, cleanup of the API
is also left for a future commit.