Use named return parameter for RunReviewTests
Originally we explicitly didn't use a named return parameter since
handling a null JSON value required extra code and so was more work to
handle. However, this is no longer the case since switching to Angular
for our frontend, so we can remove the redundancy in favor of a slight
increase in code readability.