Add app and app update download links on review page
Notice the download links are put directly in HTML anchor elements. This
works perfectly fine, but it's a little ugly since the download link is
shown on hover. This could be improved by creating a button instead of
an anchor and downloading the file through TypeScript, but that can get
a little complicated so it's left for a future change.

Closes #75