Simplify signature for DB.PublishApp()
Originally db.PublishApp accepted all of the data it needed in its
parameters as a caching optimization instead of getting it via
DB.GetSubmittedAppInfo since api.PublishApp calls DB.GetSubmittedAppInfo
itself before calling DB.PublishApp. However, this optimization is of
negligible benefit for the foreseeable future and it complicates the
database API, making it particularly annoying to test.

Fix this by changing DB.PublishApp to take only an app ID as a parameter
and retrieve the other data it needs via DB.GetSubmittedAppInfo.