Encapsulate file storage operations
App and icon files are no longer stored in local hardcoded paths in
route handlers. Instead, routes are injected with a global implementor
of data.FileStorage. For now its only implementation is
data.LocalStorage, which stores files locally under a given directory,
but it could be expanded in the future to support different cloud
storage backends.

Apart from allowing implementations to be swapped out, this opens the
door for us to start testing file I/O logic.