server: Put reading config from env behind compile flag
The current behavior for reading the server config for both the
devconsole and reposerver is:

1. Try to read config from disk
2. If that fails, read from the environment

This allows us to read the configuration from disk on a production
server while using environment variables for local development. However,
this poses a problem: if the server fails to read the config from disk
in production, it would fall back to the environment _which is most
likely empty_.

In particular, this means that if the reposerver can't read its config
file in production, it will fall back to the (empty) environment and set
its API key to an empty string. I don't think I need to explain why
that's bad.

To prevent the server admins from screwing up, this change makes both
the devconsole and reposerver read their configs from the disk by
default, hard failing otherwise. Since reading configs from the
environment is still useful for development, that behavior is gated
behind the "debug" build tag so that the builder has to go out of their
way to enable it.