console: Compare ghID against signer's in constant time
Currently the user's GitHub user ID is checked against the configured
signer's/publisher's in variable time. This does not allow for an
authorization bypass since the user can't spoof the GitHub user they're
authenticating as. However, this is an information disclosure issue for
instances of the developer portal where the signer identity isn't
already well-known. In our case this is fine because it's already
well-known who the signer is on the official instance (me), but in the
interest of protecting the confidentiality of groups running their own
instances, we should make this comparison in constant time.