repo: Set repodata slice fields to empty instead of nil
When these repository metadata struct fields are nil, they are
serialized as null JSON values on the wire. The client doesn't know how
to handle this and so sees the metadata as invalid. It only understands
empty arrays.

This used to work fine but was accidentally broken by "7eb8741: Use
`nil` for empty slice values". To prevent the same mistake from
happening again, a comment has been added to the relevant code
explaining why empty slices are used.