Fix icon file corruption when saving
Similarly to "f017d94: Fix icon hash calculation", reads of
`formIconFile` before its saving to disk (specifically, reads to
determine the image metadata) move the file reader past the beginning of
the file, causing later writes to exclude the first kilobyte or so of
the image.

Seek to the beginning of the file before saving it to ensure we save the
entire icon.