console: Decouple app and icon files in storage
An app's APKs and icon aren't inherently linked by anything present in
their file handles, so using DeleteApp() to delete both the app and its
icon is hacky at best. Decouple these so that APKs and their icons can
be deleted separately.

Note that for LocalStorage, this change leaves extraneous empty
directories on disk whenever its contained APKs and/or icon are/is
deleted. This isn't a significant issue, so it'll be addressed in a
later change.