console: Don't save files in random subfolders
Storing APK sets and app icons in folders with randomly-generated names
currently carries the consequence of leaving empty folders around when
we delete the contained files. We _could_ check for this and delete the
folder when it's empty, but we don't have much of a reason to use
subfolders anymore, so it's better if we just don't generate them in the
first place and instead randomize the file names.

The developer console doesn't inherently expect the previous file
structure, so upgrading should be seamless.