console: Sanitize versionName of uploaded apps
An application's versionName is controlled by the developer and so
should be considered untrusted data. However, the developer console
doesn't sanitize it and it is used in URL construction for publishing to
the repository server, allowing the developer to potentially modify the
path published to. Specifically, if they make their versionName of the
form "../../<arbitrary-app-id>/<version-code>/<version-name>", they can
essentially upload an update for another app.

This exploit doesn't work as-is on the developer console since the
repository server doesn't normalize URLs and so will properly return a
404 when attacked. It also doesn't work on our production server
configuration. However, with a modified configuration of a reverse proxy
performing relative URL resolution such as nginx [1], this vulnerability
would be exploitable.

To exploit this vulnerability, a developer would need to be whitelisted
in the developer console and either 1. have an existing approved app
they can upload an update for or 2. convince a publisher to publish
their app with the visibly suspicious versionName. If they already have
an app, they just need to upload an otherwise well-formed update with a
malicious versionName as described above.

The impact is a temporary DoS of the targeted app(s) since the
Accrescent client verifies app IDs, signing certificates, version codes,
and version names before attempting to install the requested app.
However, it's worth noting that this can potentially be carried out
against a large number of apps if executed correctly since there is
currently no rate limiting for uploaded app updates.

This change is a bit of a hack; a versionName with a leading "../../" is
technically valid from the Android side of things, so ideally we'd
instead move versionName out of the URL path entirely.

[1]: https://nginx.org/en/docs/http/ngx_http_core_module.html#location